tests/wpt/web-platform-tests/content-security-policy/worker-src/shared-fallback.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/testharness-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="worker-src http://{{host}}:{{ports[http][0]}} blob:; child-src 'none'">
<script>
  var url = new URL("../support/ping.js", document.baseURI).toString();
  assert_shared_worker_is_loaded(url, "Same-origin dedicated worker allowed by 'self'.");

  var b = new Blob(["onconnect = e => { e.ports[0].postMessage('ping'); }"], {type: "text/javascript"});
  var url = URL.createObjectURL(b);
  assert_shared_worker_is_loaded(url, "blob: dedicated worker allowed by 'blob:'.");
</script>
